+++ verify/async_test.go
@@ -1,11 +1,11 @@
-package f_test
+package verify_test
 
 import (
 	"context"
 	"testing"
 	"time"
 
-	"github.com/pellared/fluentassert/f"
+	"github.com/pellared/fluentassert/verify"
 )
 
 func TestPeriodic(t *testing.T) {
@@ -14,14 +14,14 @@ func TestPeriodic(t *testing.T) {
 
 	t.Run("Eventually", func(t *testing.T) {
 		t.Run("InitialPassed", func(t *testing.T) {
-			msg := f.Periodic(timeout, interval, func() f.FailureMessage {
+			msg := verify.Periodic(timeout, interval, func() verify.FailureMessage {
 				return ""
 			}).Eventually()
 			assertPassed(t, msg)
 		})
 		t.Run("SecondPassed", func(t *testing.T) {
 			shouldPass := false
-			msg := f.Periodic(timeout, interval, func() f.FailureMessage {
+			msg := verify.Periodic(timeout, interval, func() verify.FailureMessage {
 				if !shouldPass {
 					shouldPass = true // next exeucution will pass
 					return "fail"
@@ -33,14 +33,14 @@ func TestPeriodic(t *testing.T) {
 		t.Run("NeverReturned", func(t *testing.T) {
 			ch := make(chan struct{}, 1)
 			defer close(ch)
-			msg := f.Periodic(timeout, interval, func() f.FailureMessage {
+			msg := verify.Periodic(timeout, interval, func() verify.FailureMessage {
 				<-ch
 				return ""
 			}).Eventually()
 			assertFailed(t, msg, "timeout\nfunction has never returned")
 		})
 		t.Run("Failed", func(t *testing.T) {
-			msg := f.Periodic(timeout, interval, func() f.FailureMessage {
+			msg := verify.Periodic(timeout, interval, func() verify.FailureMessage {
 				return "constant failure"
 			}).Eventually()
 			assertFailed(t, msg, "timeout\nfunction always failed\nlast failure message:\nconstant failure")
@@ -49,7 +49,7 @@ func TestPeriodic(t *testing.T) {
 
 	t.Run("EventuallyContext", func(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
-			msg := f.Periodic(timeout, interval, func() f.FailureMessage {
+			msg := verify.Periodic(timeout, interval, func() verify.FailureMessage {
 				return ""
 			}).EventuallyContext(context.Background())
 			assertPassed(t, msg)
@@ -67,7 +67,7 @@ func TestAsync(t *testing.T) {
 			defer timer.Stop()
 			ticker := time.NewTicker(interval)
 			defer ticker.Stop()
-			msg := f.Async(timer.C, ticker.C, func() f.FailureMessage {
+			msg := verify.Async(timer.C, ticker.C, func() verify.FailureMessage {
 				return ""
 			}).Eventually()
 			assertPassed(t, msg)
@@ -75,7 +75,7 @@ func TestAsync(t *testing.T) {
 		t.Run("TimeoutBeforeStart", func(t *testing.T) {
 			ch := make(chan struct{})
 			close(ch)
-			msg := f.Async(ch, ch, func() f.FailureMessage {
+			msg := verify.Async(ch, ch, func() verify.FailureMessage {
 				return ""
 			}).Eventually()
 			assertFailed(t, msg, "timeout\nfunction has never returned")
@@ -88,7 +88,7 @@ func TestAsync(t *testing.T) {
 			defer timer.Stop()
 			ticker := time.NewTicker(interval)
 			defer ticker.Stop()
-			msg := f.Async(timer.C, ticker.C, func() f.FailureMessage {
+			msg := verify.Async(timer.C, ticker.C, func() verify.FailureMessage {
 				return ""
 			}).EventuallyContext(context.Background())
 			assertPassed(t, msg)
@@ -97,7 +97,7 @@ func TestAsync(t *testing.T) {
 			ctx, cancel := context.WithCancel(context.Background())
 			cancel()
 			ch := make(chan struct{})
-			msg := f.Async(ch, ch, func() f.FailureMessage {
+			msg := verify.Async(ch, ch, func() verify.FailureMessage {
 				return ""
 			}).EventuallyContext(ctx)
 			assertFailed(t, msg, "context canceled\nfunction has never returned")
@@ -106,7 +106,7 @@ func TestAsync(t *testing.T) {
 			ctx, cancel := context.WithTimeout(context.Background(), timeout)
 			defer cancel()
 			ch := make(chan struct{})
-			msg := f.Async(ch, ch, func() f.FailureMessage {
+			msg := verify.Async(ch, ch, func() verify.FailureMessage {
 				return "constant failure"
 			}).EventuallyContext(ctx)
 			assertFailed(t, msg, "context deadline exceeded\nfunction always failed\nlast failure message:\nconstant failure")