+++ verify/doc.go
@@ -1,8 +1,8 @@
-// Package f contains the most useful type-safe fluent assertion functions.
+// Package verify contains the most useful type-safe fluent assertion functions.
 //
 // It also provides the FailureMessage type which you can use
 // for creating your own fluent assertions.
 //
 // At last, you may embed a Fluent* type
 // to extend it with additional assertion functions.
-package f
+package verify