+++ verify/failmsg_test.go
@@ -1,22 +1,22 @@
-package f_test
+package verify_test
 
 import (
 	"testing"
 
-	"github.com/pellared/fluentassert/f"
+	"github.com/pellared/fluentassert/verify"
 )
 
 func TestFailureMessage(t *testing.T) {
 	t.Run("Assert", func(t *testing.T) {
 		t.Run("Empty", func(t *testing.T) {
 			mock := &errorMock{}
-			got := f.FailureMessage("").Assert(mock)
+			got := verify.FailureMessage("").Assert(mock)
 			assertTrue(t, got)
 			assertEqual(t, mock, &errorMock{})
 		})
 		t.Run("NoArgs", func(t *testing.T) {
 			mock := &errorMock{}
-			got := f.FailureMessage("failed").Assert(mock)
+			got := verify.FailureMessage("failed").Assert(mock)
 			assertFalse(t, got)
 			assertEqual(t, mock, &errorMock{
 				HelperCalled: true,
@@ -26,7 +26,7 @@ func TestFailureMessage(t *testing.T) {
 		})
 		t.Run("WithArgs", func(t *testing.T) {
 			mock := &errorMock{}
-			got := f.FailureMessage("failed").Assert(mock, "arg1", 2)
+			got := verify.FailureMessage("failed").Assert(mock, "arg1", 2)
 			assertFalse(t, got)
 			assertEqual(t, mock, &errorMock{
 				HelperCalled: true,
@@ -39,13 +39,13 @@ func TestFailureMessage(t *testing.T) {
 	t.Run("Require", func(t *testing.T) {
 		t.Run("Empty", func(t *testing.T) {
 			mock := &fatalMock{}
-			got := f.FailureMessage("").Require(mock)
+			got := verify.FailureMessage("").Require(mock)
 			assertTrue(t, got)
 			assertEqual(t, mock, &fatalMock{})
 		})
 		t.Run("NoArgs", func(t *testing.T) {
 			mock := &fatalMock{}
-			got := f.FailureMessage("failed").Require(mock)
+			got := verify.FailureMessage("failed").Require(mock)
 			assertFalse(t, got)
 			assertEqual(t, mock, &fatalMock{
 				HelperCalled: true,
@@ -55,7 +55,7 @@ func TestFailureMessage(t *testing.T) {
 		})
 		t.Run("WithArgs", func(t *testing.T) {
 			mock := &fatalMock{}
-			got := f.FailureMessage("failed").Require(mock, "arg1", 2)
+			got := verify.FailureMessage("failed").Require(mock, "arg1", 2)
 			assertFalse(t, got)
 			assertEqual(t, mock, &fatalMock{
 				HelperCalled: true,
@@ -68,13 +68,13 @@ func TestFailureMessage(t *testing.T) {
 	t.Run("Assertf", func(t *testing.T) {
 		t.Run("Empty", func(t *testing.T) {
 			mock := &errorfMock{}
-			got := f.FailureMessage("").Assertf(mock, "should pass")
+			got := verify.FailureMessage("").Assertf(mock, "should pass")
 			assertTrue(t, got)
 			assertEqual(t, mock, &errorfMock{})
 		})
 		t.Run("NoArgs", func(t *testing.T) {
 			mock := &errorfMock{}
-			got := f.FailureMessage("failed").Assertf(mock, "should pass")
+			got := verify.FailureMessage("failed").Assertf(mock, "should pass")
 			assertFalse(t, got)
 			assertEqual(t, mock, &errorfMock{
 				HelperCalled: true,
@@ -85,7 +85,7 @@ func TestFailureMessage(t *testing.T) {
 		})
 		t.Run("WithArgs", func(t *testing.T) {
 			mock := &errorfMock{}
-			got := f.FailureMessage("failed").Assertf(mock, "should work %d", 1)
+			got := verify.FailureMessage("failed").Assertf(mock, "should work %d", 1)
 			assertFalse(t, got)
 			assertEqual(t, mock, &errorfMock{
 				HelperCalled: true,
@@ -99,13 +99,13 @@ func TestFailureMessage(t *testing.T) {
 	t.Run("Requiref", func(t *testing.T) {
 		t.Run("Empty", func(t *testing.T) {
 			mock := &fatalfMock{}
-			got := f.FailureMessage("").Requiref(mock, "should pass")
+			got := verify.FailureMessage("").Requiref(mock, "should pass")
 			assertTrue(t, got)
 			assertEqual(t, mock, &fatalfMock{})
 		})
 		t.Run("NoArgs", func(t *testing.T) {
 			mock := &fatalfMock{}
-			got := f.FailureMessage("failed").Requiref(mock, "should pass")
+			got := verify.FailureMessage("failed").Requiref(mock, "should pass")
 			assertFalse(t, got)
 			assertEqual(t, mock, &fatalfMock{
 				HelperCalled: true,
@@ -116,7 +116,7 @@ func TestFailureMessage(t *testing.T) {
 		})
 		t.Run("WithArgs", func(t *testing.T) {
 			mock := &fatalfMock{}
-			got := f.FailureMessage("failed").Requiref(mock, "should work %d", 1)
+			got := verify.FailureMessage("failed").Requiref(mock, "should work %d", 1)
 			assertFalse(t, got)
 			assertEqual(t, mock, &fatalfMock{
 				HelperCalled: true,
@@ -129,19 +129,19 @@ func TestFailureMessage(t *testing.T) {
 
 	t.Run("Merge", func(t *testing.T) {
 		t.Run("BothEmpty", func(t *testing.T) {
-			var first, second f.FailureMessage
+			var first, second verify.FailureMessage
 			got := first.Merge("assertion", second)
 			assertPassed(t, got)
 		})
 		t.Run("ArgIsNotEmpty", func(t *testing.T) {
-			var msg f.FailureMessage
-			got := msg.Merge("assertion", f.FailureMessage("failure"))
+			var msg verify.FailureMessage
+			got := msg.Merge("assertion", verify.FailureMessage("failure"))
 			assertFailed(t, got, "assertion\nfailure")
 		})
 		t.Run("NoneIsEmpty", func(t *testing.T) {
-			var msg f.FailureMessage
-			msg = msg.Merge("first", f.FailureMessage("error"))
-			got := msg.Merge("second", f.FailureMessage("failure"))
+			var msg verify.FailureMessage
+			msg = msg.Merge("first", verify.FailureMessage("error"))
+			got := msg.Merge("second", verify.FailureMessage("failure"))
 			assertFailed(t, got, "first\nerror\n\nsecond\nfailure")
 		})
 	})