+++ verify/obj_test.go
@@ -1,9 +1,9 @@
-package f_test
+package verify_test
 
 import (
 	"testing"
 
-	"github.com/pellared/fluentassert/f"
+	"github.com/pellared/fluentassert/verify"
 )
 
 func TestObj(t *testing.T) {
@@ -17,18 +17,18 @@ func TestObj(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
 			got := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
-			msg := f.Obj(got).DeepEqual(want)
+			msg := verify.Obj(got).DeepEqual(want)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
 			got := A{Str: "wrong", Bool: true, Slice: []int{1, 3}}
-			msg := f.Obj(got).DeepEqual(want)
+			msg := verify.Obj(got).DeepEqual(want)
 			assertFailed(t, msg, "mismatch (-want +got):\n")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := f.Obj(got).DeepEqual(nil)
+			msg := verify.Obj(got).DeepEqual(nil)
 			assertPassed(t, msg)
 		})
 	})
@@ -37,18 +37,18 @@ func TestObj(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
 			got := A{Str: "wrong", Bool: true, Slice: []int{1, 3}}
-			msg := f.Obj(got).NotDeepEqual(want)
+			msg := verify.Obj(got).NotDeepEqual(want)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
 			got := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
-			msg := f.Obj(got).NotDeepEqual(want)
+			msg := verify.Obj(got).NotDeepEqual(want)
 			assertFailed(t, msg, "the objects are equal")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := f.Obj(got).NotDeepEqual(nil)
+			msg := verify.Obj(got).NotDeepEqual(nil)
 			assertFailed(t, msg, "the objects are equal")
 		})
 	})
@@ -56,17 +56,17 @@ func TestObj(t *testing.T) {
 	t.Run("Zero", func(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			got := A{}
-			msg := f.Obj(got).Zero()
+			msg := verify.Obj(got).Zero()
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			got := A{Str: "wrong"}
-			msg := f.Obj(got).Zero()
+			msg := verify.Obj(got).Zero()
 			assertFailed(t, msg, "not a zero value (-want +got):\n")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := f.Obj(got).Zero()
+			msg := verify.Obj(got).Zero()
 			assertPassed(t, msg)
 		})
 	})
@@ -74,34 +74,34 @@ func TestObj(t *testing.T) {
 	t.Run("NonZero", func(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			got := A{Str: "string"}
-			msg := f.Obj(got).NonZero()
+			msg := verify.Obj(got).NonZero()
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			got := A{}
-			msg := f.Obj(got).NonZero()
+			msg := verify.Obj(got).NonZero()
 			assertFailed(t, msg, "a zero value")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := f.Obj(got).NonZero()
+			msg := verify.Obj(got).NonZero()
 			assertFailed(t, msg, "a zero value")
 		})
 	})
 
 	t.Run("Check", func(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
-			fn := func(x A) f.FailureMessage {
+			fn := func(x A) verify.FailureMessage {
 				return ""
 			}
-			msg := f.Obj(A{}).Check(fn)
+			msg := verify.Obj(A{}).Check(fn)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
-			fn := func(x A) f.FailureMessage {
+			fn := func(x A) verify.FailureMessage {
 				return "failure"
 			}
-			msg := f.Obj(A{}).Check(fn)
+			msg := verify.Obj(A{}).Check(fn)
 			assertFailed(t, msg, "failure")
 		})
 	})
@@ -111,14 +111,14 @@ func TestObj(t *testing.T) {
 			pred := func(x A) bool {
 				return true
 			}
-			msg := f.Obj(A{}).Should(pred)
+			msg := verify.Obj(A{}).Should(pred)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			pred := func(x A) bool {
 				return false
 			}
-			msg := f.Obj(A{}).Should(pred)
+			msg := verify.Obj(A{}).Should(pred)
 			assertFailed(t, msg, "object does not meet the predicate criteria")
 		})
 	})
@@ -128,14 +128,14 @@ func TestObj(t *testing.T) {
 			pred := func(x A) bool {
 				return false
 			}
-			msg := f.Obj(A{}).ShouldNot(pred)
+			msg := verify.Obj(A{}).ShouldNot(pred)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			pred := func(x A) bool {
 				return true
 			}
-			msg := f.Obj(A{}).ShouldNot(pred)
+			msg := verify.Obj(A{}).ShouldNot(pred)
 			assertFailed(t, msg, "object meets the predicate criteria")
 		})
 	})